Added `manage.py` in example directory
Hi @nitely, this is work in progress to discuss about these changes, we
may rebase later if you find this commits usefull.

I refactored the `example` in order to get used by users who wants to
test rapidly the Spirit app, this is inspired by `django-allauth` package.

At the moment it doesn't do any test on it, but we could run tests with
these settings, urls and wsgi files too.